requirements.txt
Django==2.0.3 
django_extensions
django-import-export
django-suit